Revert "float -> double"
This reverts commit f010c205f47e68fa96dc71f65fe64b6a054b1d49.

+++ Engine/FrameTimer.h
@@ -5,7 +5,7 @@ class FrameTimer
 {
 public:
 	FrameTimer();
-	double Mark();
+	float Mark();
 private:
 	std::chrono::steady_clock::time_point last;
 };
